Refactor auth components. Move logic into custom hooks
