+++ frontend/src/component/ResetPassword.js
@@ -1,43 +1,12 @@
-import React, { useState } from 'react';
-import { Link, useParams } from 'react-router-dom';
-import { Container, Box, Typography, Grid, TextField, Button } from '@mui/material';
-import useAlert from '../hook/useAlert';
+import React from 'react';
+import { Container, Box, Typography, Grid, Button } from '@mui/material';
 import StyledTextField from './StyledTextField';
+import useResetPassword from '../hook/useResetPassword';
 
 const ResetPassword = () => {
-  const { resetCode } = useParams();
-  const { handleAlertChange, AlertComponent } = useAlert();
-  const [password, setPassword] = useState('');
 
-  const handlePasswordChange = (event) => {
-    setPassword(event.target.value);
-  };
+  const { handlePasswordChange, handleResetPassword } = useResetPassword();
 
-  const handleResetPassword = async () => {
-    // check that user typed a password
-    try {
-      if (!password) throw new Error('Please fill in the password');
-      const resetData = {
-        password
-      };
-      console.log(JSON.stringify(resetData));
-      const response = await fetch(`http://localhost:4200/auth/reset-password/${resetCode}`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(resetData)
-      });
-      const responseData = await response.json();
-      if (response.ok) {
-        handleAlertChange("Password reset successful. Try logging in now.", "info");
-        return;
-      }
-      handleAlertChange(`${responseData.error}`, "error");
-    } catch (error) {
-      handleAlertChange(`${error}`, "error");
-    }
-  };
   return (
     <Container>
       <Box
@@ -69,24 +38,23 @@ const ResetPassword = () => {
               type='password'
               margin="normal"
               variant='filled'
-              value={password}
+              // value={password}
               onChange={handlePasswordChange}
             />
           </Grid>
         </Grid>
-          <Button variant="contained" color="primary" onClick={handleResetPassword} mt={2} sx={{
-            width: '241px',
-            height: '53px',
-            marginTop: '15px',
-            padding: '16px 32px 16px 32px',
-            borderRadius: '100px',
-            gap: '20px',
-            textTransform: 'none',
-          }}>
-            Save
-          </Button>
+        <Button variant="contained" color="primary" onClick={handleResetPassword} mt={2} sx={{
+          width: '241px',
+          height: '53px',
+          marginTop: '15px',
+          padding: '16px 32px 16px 32px',
+          borderRadius: '100px',
+          gap: '20px',
+          textTransform: 'none',
+        }}>
+          Save
+        </Button>
       </Box>
-      { AlertComponent }
     </Container>
   );
 };
